requirements.txt
hidapi==0.14.0.post2 ; python_version >= "3.10" and python_version < "4.0"
psutil==6.1.0 ; python_version >= "3.10" and python_version < "4.0"
typer==0.12.5 ; python_version >= "3.10" and python_version < "4.0"
